FemtoVG renderer: Separate the font database from the font cache
This is done in preparation for moving the database to core.
