+++ internal/renderers/femtovg/sharedfontdb/DejaVuSans.ttf
